# application/services/custom_services/gporder/src/models/entitymodels/item.ts
export interface item 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   type: { type: String },
   name: { type: String },
   description: { type: String },
   price: { type: String },
   price_currency_type: { type: String },
   cost: { type: String },
   cost_currency_type: { type: String },
   gephistoryid: { type: String }
}